refactor(api): split BookingService into query/creation/lifecycle (#713)
Decompose the 798-line BookingService god-class into three single-
responsibility services that share the same injected repos via the
composition root:

- BookingQueryService    (booking-query.ts, 8 read/list methods + BookingWithOperator)
- BookingCreationService (booking-creation.ts, create + private submitInTx)
- BookingLifecycleService(booking-lifecycle.ts, updateStatus/cancel/substitute
                          /sameAcrissClass/findSubstitutionCandidates)

booking.ts is now a thin composite facade (190 lines) that wires the three
and delegates every public method, so routes and tests keep a single
BookingService dependency. Shared time/price helpers (rentalDays, MS_*)
live in booking-pricing-helpers.ts. Behavior-preserving: no logic changes,
the full api unit (1357) + integration (234) suites pass unchanged.

Closes #713
