+++ packages/api/src/services/booking-creation.ts
@@ -1,0 +1,334 @@
+import type { AddOnSnapshot, InsuranceSnapshot } from '@kuruma/shared/db/schema'
+import { calculateBookingPrice } from '@kuruma/shared/lib/pricing'
+import { checkRentalRules } from '@kuruma/shared/lib/rental-rules'
+import { generateBookingCode } from '../lib/booking-code'
+import { type CallerContext, SYSTEM_CONTEXT } from '../middleware/auth'
+import { BOOKING_CODE_CONSTRAINT, PG_ERROR, pgConstraintName, pgErrorCode } from '../pg-errors'
+import type {
+  BookingRepository,
+  RunInTransaction,
+  TransactionRepos,
+  UserRepository,
+} from '../repositories/types'
+import type { BookingPostCommitDispatcher } from './booking-post-commit-dispatcher'
+import { MS_PER_MINUTE, rentalDays } from './booking-pricing-helpers'
+import type {
+  BookingVerificationGate,
+  CreateBookingInput,
+  CreateBookingResult,
+} from './booking-types'
+
+// Location-only turnaround fallback when a pickup location has no value set
+// (§5.3, proposal §9 item 20). The legacy 60-min vehicle buffer is GONE — a
+// 60-min result is a regression the turnaround test pins against.
+const DEFAULT_TURNAROUND_MINUTES = 2880 // 48h
+// A booking_code collision is ~2^-40 per attempt; a few retries is plenty.
+const MAX_BOOKING_CODE_ATTEMPTS = 3
+// #613: version of the liability-disclaimer (免责声明) wording a renter agreed to,
+// stamped onto the booking. Bump when the localized terms text changes materially
+// (the renter-facing copy lives in web i18n; this is the server-authoritative tag).
+export const DISCLAIMER_TERMS_VERSION = '2026-06-13'
+
+/**
+ * Creation side of the booking domain (#713 split of the BookingService
+ * god-class). Owns `create` (idempotency + verification gate + bounded
+ * collision-retry) and the private `submitInTx` decision-and-record. All repos
+ * are injected; the composition root shares the same instances across the
+ * query/creation/lifecycle services.
+ */
+export class BookingCreationService {
+  constructor(
+    private readonly bookingRepo: BookingRepository,
+    private readonly runInTransaction: RunInTransaction,
+    private readonly userRepo?: UserRepository,
+    // Single post-commit seam (#393, TODO #300): ensureThread + notifications,
+    // each caught-and-logged. Replaces the inline ensureThread calls.
+    private readonly postCommit?: BookingPostCommitDispatcher,
+    // Injectable so the collision-retry path is deterministically testable.
+    private readonly generateCode: () => string = generateBookingCode,
+    // #459: optional document-verification gate, wired only when the
+    // REQUIRE_DOCUMENT_VERIFICATION flag is on (see index.ts composition root).
+    private readonly verificationGate?: BookingVerificationGate,
+  ) {}
+
+  async create(
+    ctx: CallerContext,
+    input: CreateBookingInput,
+    now: Date = new Date(),
+  ): Promise<CreateBookingResult> {
+    // Staff/admin override path: validate the target renterId exists and is a
+    // RENTER — prevents bookings attached to other staff/admin accounts and
+    // surfaces FK failures as a friendly 400 rather than a generic 500.
+    if (input.renterId !== ctx.userId && this.userRepo) {
+      const [renter] = await this.userRepo.findByIds([input.renterId])
+      if (!renter) {
+        return { ok: false, status: 400, error: 'Renter not found' }
+      }
+      if ((renter.role ?? 'RENTER') !== 'RENTER') {
+        return { ok: false, status: 400, error: 'Target user is not a renter' }
+      }
+    }
+
+    // Idempotency: if a key was provided, replay an existing booking first.
+    if (input.idempotencyKey) {
+      const existing = await this.bookingRepo.findByIdempotencyKey(ctx, input.idempotencyKey)
+      if (existing) {
+        await this.postCommit?.run(ctx, existing)
+        return { ok: true, booking: existing, status: 200 }
+      }
+    }
+
+    // Document-verification gate (#459). When wired (REQUIRE_DOCUMENT_VERIFICATION
+    // on), an unverified renter is blocked here — before any booking work. Runs
+    // AFTER idempotency replay so a previously-accepted booking stays replayable
+    // even if the renter's document later expires.
+    if (this.verificationGate) {
+      const gate = await this.verificationGate(ctx, input)
+      if (!gate.ok) return gate
+    }
+
+    // The whole submit is ONE transaction (proposal §4): resolve + price +
+    // snapshot + insert booking + append BOOKING_CREATED, atomically. The
+    // booking_code is generated inside the tx, so a UNIQUE clash re-runs the
+    // entire atomic insert cleanly (§5.4) — bounded retry around the tx.
+    let lastErr: unknown
+    for (let attempt = 0; attempt < MAX_BOOKING_CODE_ATTEMPTS; attempt++) {
+      try {
+        const result = await this.runInTransaction((repos) =>
+          this.submitInTx(ctx, input, now, repos),
+        )
+        if (!result.ok) return result // domain validation failure — never retried
+        // Post-commit side effects (#335 thread + #393 notifications) — never in
+        // the tx; awaited, caught-and-logged in the dispatcher (booking authoritative).
+        await this.postCommit?.run(ctx, result.booking)
+        return { ok: true, booking: result.booking }
+      } catch (err) {
+        const code = pgErrorCode(err)
+        if (
+          code === PG_ERROR.UNIQUE_VIOLATION &&
+          pgConstraintName(err) === BOOKING_CODE_CONSTRAINT
+        ) {
+          lastErr = err // regenerate + retry the whole atomic insert
+          continue
+        }
+        if (code === PG_ERROR.EXCLUSION_VIOLATION) {
+          return {
+            ok: false,
+            status: 409,
+            error: 'Vehicle is already booked for the requested time range',
+          }
+        }
+        // Idempotency-key race: a concurrent replay won — return its booking.
+        if (code === PG_ERROR.UNIQUE_VIOLATION && input.idempotencyKey) {
+          const existing = await this.bookingRepo.findByIdempotencyKey(ctx, input.idempotencyKey)
+          if (existing) {
+            await this.postCommit?.run(ctx, existing)
+            return { ok: true, booking: existing, status: 200 }
+          }
+        }
+        throw err
+      }
+    }
+    // Exhausted the bounded retries — surface as a 500 at the route boundary.
+    throw lastErr
+  }
+
+  // The atomic decision-and-record (FC/IS: the decision; ensureThread is the
+  // imperative shell). Resolution reads run as SYSTEM_CONTEXT — these are
+  // internal pricing/snapshot reads, and insurance/fee reads reject RENTER
+  // callers by design; the booking's tenant is derived from the vehicle, not
+  // the caller. The booking + event WRITES use the caller's ctx (renter-own /
+  // operator scope is enforced at the repo).
+  private async submitInTx(
+    ctx: CallerContext,
+    input: CreateBookingInput,
+    now: Date,
+    repos: TransactionRepos,
+  ): Promise<CreateBookingResult> {
+    const vehicle = await repos.vehicleRepo.findById(SYSTEM_CONTEXT, input.requestedVehicleId)
+    if (!vehicle) {
+      return { ok: false, status: 400, error: 'Vehicle not found' }
+    }
+    if (vehicle.status !== 'AVAILABLE') {
+      return { ok: false, status: 400, error: 'Vehicle is not available' }
+    }
+    if (!vehicle.classId) {
+      return { ok: false, status: 400, error: 'Vehicle has no class' }
+    }
+    const operatorId = vehicle.operatorId
+    const classId = vehicle.classId
+    const assignedVehicleId = vehicle.id // server-derived = requested at submit
+
+    const ruleCheck = checkRentalRules(
+      {
+        minRentalHours: vehicle.minRentalHours,
+        maxRentalHours: vehicle.maxRentalHours,
+        advanceBookingHours: input.source === 'MANUAL' ? null : vehicle.advanceBookingHours,
+      },
+      input.startAt,
+      input.endAt,
+      now,
+    )
+    if (!ruleCheck.ok) {
+      return {
+        ok: false,
+        status: 400,
+        error: 'Booking violates a rental rule on this vehicle',
+        code: ruleCheck.code,
+        details: { required: ruleCheck.required, actual: ruleCheck.actual },
+      }
+    }
+
+    // The car physically lives at its own storefront, so a booking can only pick
+    // it up there. Without this, a forged body could pair a vehicle with another
+    // of the operator's locations — the operator check below passes (same tenant)
+    // but the stamped pickup/turnaround would lie about where the car is (#392).
+    if (vehicle.pickupLocationId !== input.pickupLocationId) {
+      return { ok: false, status: 400, error: 'Pickup location does not match the vehicle' }
+    }
+
+    // Turnaround is location-only (§5.3): pickup location's default, 48h fallback.
+    const pickup = await repos.locationRepo.findById(SYSTEM_CONTEXT, input.pickupLocationId)
+    if (!pickup || pickup.operatorId !== operatorId) {
+      return { ok: false, status: 400, error: 'Pickup location is not available' }
+    }
+    if (input.dropoffLocationId !== input.pickupLocationId) {
+      const dropoff = await repos.locationRepo.findById(SYSTEM_CONTEXT, input.dropoffLocationId)
+      if (!dropoff || dropoff.operatorId !== operatorId) {
+        return { ok: false, status: 400, error: 'Dropoff location is not available' }
+      }
+    }
+    const turnaroundMinutes = pickup.defaultTurnaroundMinutes ?? DEFAULT_TURNAROUND_MINUTES
+    const effectiveEndAt = new Date(input.endAt.getTime() + turnaroundMinutes * MS_PER_MINUTE)
+
+    // Price off the ASSIGNED vehicle's rates — never class-level (#406), never
+    // client-supplied (#74), non-null on every submit (#429 backfill guard).
+    const pricing = calculateBookingPrice(
+      { dailyRateJpy: vehicle.dailyRateJpy, hourlyRateJpy: vehicle.hourlyRateJpy },
+      input.startAt,
+      input.endAt,
+    )
+    if (!pricing.ok) {
+      return {
+        ok: false,
+        status: 400,
+        error:
+          pricing.code === 'NO_RATES_SET'
+            ? 'No daily or hourly rate configured'
+            : 'Invalid booking duration',
+        code: pricing.code,
+      }
+    }
+    let totalPrice = pricing.totalPriceJpy
+
+    // Selected insurance: snapshot the chosen ACTIVE option from THIS operator.
+    let insuranceOptionId: string | null = null
+    let insuranceSnapshot: InsuranceSnapshot | null = null
+    if (input.insuranceOptionId) {
+      const opt = await repos.insuranceOptionRepo.findById(SYSTEM_CONTEXT, input.insuranceOptionId)
+      if (!opt || opt.operatorId !== operatorId || opt.status !== 'ACTIVE') {
+        return { ok: false, status: 400, error: 'Insurance option is not available' }
+      }
+      insuranceOptionId = opt.id
+      insuranceSnapshot = {
+        insuranceOptionId: opt.id,
+        name: opt.name,
+        dailyPriceJpy: opt.dailyPriceJpy,
+        deductibleJpy: opt.deductibleJpy,
+      }
+      totalPrice += opt.dailyPriceJpy * rentalDays(input.startAt, input.endAt)
+    }
+
+    // Fee snapshot: this operator's ACTIVE fees that are operator-wide or match
+    // the booking's class. Locks the rate at booking time (§6.2, informational).
+    const fees = await repos.feeScheduleRepo.findAll(SYSTEM_CONTEXT, {
+      operatorId,
+      status: 'ACTIVE',
+    })
+    const feeSnapshot = fees
+      .filter((f) => f.vehicleClassId === null || f.vehicleClassId === classId)
+      .map((f) => ({
+        feeType: f.feeType,
+        unit: f.unit,
+        amountJpy: f.amountJpy,
+        vehicleClassId: f.vehicleClassId,
+      }))
+
+    // Selected paid add-ons (#460): each flat priceJpy is added to totalPrice and
+    // snapshotted. One ACTIVE-only, this-operator read validates membership +
+    // tenant + active in a single query (a foreign/archived id is simply absent
+    // -> 400). De-dup so a repeated id is charged once (quantity is out of MVP).
+    const addOnSnapshot: AddOnSnapshot[] = []
+    if (input.addOnIds.length > 0) {
+      const available = await repos.addOnRepo.findActiveByOperator(operatorId)
+      const availableById = new Map(available.map((a) => [a.id, a]))
+      for (const addOnId of new Set(input.addOnIds)) {
+        const addOn = availableById.get(addOnId)
+        if (!addOn) return { ok: false, status: 400, error: 'Add-on is not available' }
+        addOnSnapshot.push({ addOnId: addOn.id, name: addOn.name, priceJpy: addOn.priceJpy })
+        totalPrice += addOn.priceJpy
+      }
+    }
+
+    const bookingCode = this.generateCode()
+
+    // Insert FIRST: the exclusion / unique constraints fire here, BEFORE the
+    // event append, so a rolled-back insert leaves no orphan event (atomicity).
+    const booking = await repos.bookingRepo.create(ctx, {
+      operatorId,
+      renterId: input.renterId,
+      classId,
+      requestedVehicleId: input.requestedVehicleId,
+      assignedVehicleId,
+      pickupLocationId: input.pickupLocationId,
+      dropoffLocationId: input.dropoffLocationId,
+      startAt: input.startAt,
+      endAt: input.endAt,
+      effectiveEndAt,
+      status: 'CONFIRMED',
+      source: input.source,
+      // #463: server-derived. Every pre-demo submit is SPECIFIC; CLASS_COMBO is #464.
+      fulfillmentMode: 'SPECIFIC',
+      bookingCode,
+      insuranceOptionId,
+      insuranceSnapshot,
+      feeSnapshot,
+      addOnSnapshot,
+      externalId: input.externalId ?? null,
+      notes: input.notes ?? null,
+      totalPrice,
+      cancellationFee: null,
+      cancelledAt: null,
+      idempotencyKey: input.idempotencyKey ?? null,
+      // #613: stamp the liability-disclaimer consent server-side (never trust a
+      // client timestamp). Set only when the renter accepted; null for staff/
+      // manual bookings (the route exempts non-renter callers from the gate).
+      disclaimerAcknowledgedAt: input.disclaimerAccepted ? now : null,
+      disclaimerTermsVersion: input.disclaimerAccepted ? DISCLAIMER_TERMS_VERSION : null,
+    })
+
+    await repos.bookingEventRepo.append(ctx, {
+      bookingId: booking.id,
+      type: 'BOOKING_CREATED',
+      // The actor is who PERFORMED the booking (the authed caller), not the
+      // subject: a staff/manual booking records the staff member, not the
+      // customer. The renter lives in the booking row + payload. Mirrors
+      // VEHICLE_SUBSTITUTED, which also uses ctx.userId.
+      actorId: ctx.userId,
+      payload: {
+        requestedVehicleId: input.requestedVehicleId,
+        assignedVehicleId,
+        classId,
+        fulfillmentMode: 'SPECIFIC', // #463: mirror the booking's discriminator in the audit snapshot
+        startAt: input.startAt.toISOString(),
+        endAt: input.endAt.toISOString(),
+        totalPrice,
+        insuranceSnapshot,
+        feeSnapshot,
+        addOnSnapshot,
+      },
+    })
+
+    return { ok: true, booking }
+  }
+}
